Close SubprocessStreams in vdg TRIM_READS. (#781)
* Fix VDJ trim_reads so it actually works.

Before, trim_reads was attempting to read its own _stdout file.  This
should not, in general, be expected to work.  Even if the file exists
(which on for example LSF it does not, because LSF buffers the output
until the job completes), there's absolutely no guarantee that it's been
flushed when you go to read it.  And it's a shared file, so lots of
other things can and do write to it.

This fixes things by instead redirecting the output from cutadapt to a
file in the stage's files directory and then reading that file.

This should fix JIRA CSI-658.

* Close SubprocessStreams in vdg TRIM_READS.

Also, before waiting on the subprocess for SubprocessStream, kill it to
make sure we're not waiting forever.