Fix SubprocessStream write streams.
In a read stream, when the stream is closed the subprocess can and must be
killed if it's still running, since otherwise it might hang waiting for something to
consume its output.  However, in a write stream the process must be allowed to
complete normally after its input pipe has been closed.