Make extract_reads faster (#899)
* Remove unnessessary split for detect_chemistry.

* Leave barcode whitelist as a set where possible.

cr_utils.load_barcode_whitelist converts the list to a set to verify
uniqueness.  Allow it to return that set in cases where order is not
required so the conversion to set isn't duplicated.

Pass the barcode_whitelist to infer_barcode_reverse_complement as a set,
rather than as a list, since infer_barcode_reverse_complement is
checking lots of strings against the whitelist.

Add context manager methods to FastqReader so it can be used in a with
statement.

* Cythonize lib/python/cellranger/report.pyx

Make the report metrics cache faster for sequence types.