Replace Brute-Force MLE Estimation with Analytical Solution (#906)
Inside a bootstrapping process, this section of the code tries to find the MLE for the `N` parameter of a binomial given fixed `K`, `p`.  It does this by calculating the binomial pmf for every value of N from 0 to the number of observed cells in an experiment, and then identifying the highest value.

This brute force procedure is obfuscating.  Strictly speaking, the MLE of this model isn't identifiable, because given a K and p, the pmf for `N = k/p` and `N= k/p - 1` have identical values (e.g. `dbinom(60, 199, .3) == dbinom(60, 200, .3))`.  However, one can obtain an identifiable solution that is an analytical maximum by using a continuous analog of the discrete PMF.  Most easily, a binomial random variable can be approximated as a `Normal(N*p, N*p*(1-p))`random variable, and one can then analytically solve for the MLE and obtain the following equation for it:

    N = (-p + p^2 + p*sqrt(1 + 4*k^2 -2*p + p^2)) / (2 * p^2)

Noting that the number of observed two-genome cells (`K`) is much higher than a probability `p`, e.g. ` 0 < p < 1 && p << K` this can be simplified to `N=k/p` by dropping the smaller terms.

This estimate `N=k/p` is particularly nice not only because it is on the likelihood ridge of the non-identifiable MLE, but also because it is the value of the method of moments estimate, as `E[K] = N*p -> N = K/p`, so this analytical solution can either be viewed as a MLE or a MOM estimate but it has the same value.

Although this brute force operation is an expensive `O(# Cells)` operation, this change is for clarity rather than speed, as the doublet estimation both only applies to experiments with multiple genomes and is presumably a small percentage of the overall processing time even in that case.

In summary: N = k / p is the correct answer.

Following review, a few unit tests have also been added.