Fix IRLBA Partial SVD Code (#1022)
The IRLB code was copied from to python as a simplified/reduced feature version
of the R/C code.  Although this code generally behaved, we found that it
occasionally gave absurd results, such as saying certain eigenvectors explained
proportions of variance that were ~1e70.  Tracking this down identified two issues.

1) The code could allocate a working space larger than what is mathematically
possible, leading to absurd results being returned with no indication they were
abnormal.

2) The centering was not handled robustly.  An omission in porting the code
from R to python was typically irrelevant if the center value was near
the column mean, but could have very significant effects under some circumstances.
To correct this, the code to robustly handle the centering was ported over from R.

To help aid and ensure code quality, multiple tests and comments are part of this commit,
including tests to ensure the pathological matrices are handled appropriately going
forward.

A curiosity observed in both the R package and the python package while working with
this code was that the `irlb` function call was always more accurate with a
dense matrix than with a sparse matrix.  The difference is within numerical
accuracy (<1e-5) but unexplained.