PCA Fixes and Refactoring (#1143)
* Fix IRLBA for np.matrix types

The `mult` function in IRLBA was mislabelled, and would return different results depending on if a dense or sparse or dense matrix was passed as A, and on whether or not the dense matrix was a numpy array or a numpy matrix.  This cleans up the code so that it will work with a numpy matrix and has correct comments for documentation.

* Add Explanatory Comment

* PCA Fixes and Refactoring

*Changes to IRLB Algorithm*

In writing unit tests for and resolving issues in the PCA, I found some more issues with the IRLB code.  Namely, the `mult` function for matrix multiplication within it did not behave as claimed, but under the hood was behaving differently for sparse vs. dense matrices, and only worked for numpy.array and not numpy.matrix despite explicitly stating otherwise.

This commit is off-target for the sprint as I don't have reason to believe this issue has created any problems to date since we consistenly used sparse matrices and so haven't encountered these problems.  But, it took me so long to figure out what was going on here, I figured I might as well correct it and add more tests around it to verify before we run into it any problems from this later on.

*PCA Changes*

We've had a number of issues with the PCA of late, and this commit does a large refactoring to clean-up/clarify the code, better handle edge cases, and introduce tests of the PCA functions.  These changes can affect the programs output, and they also catch what I believe is a so far undiscovered bug.

*Changes where the thresholding to the matrix is applied* - We remove all rows/columns with a count below `min_count_threshold` before doing the PCA.  However, we randomly sample a subset of barcodes, and selected features based on their ranked dispersion, before then.  As a result, even if the full matrix was something we could perform PCA on, we might get a subsample where all the rows/columns are beneath the threshold and have to error out.  To avoid this, we threshold prior to doing the subsetting (which is one or both of random barcodes and a subset of features based on their ranked normalized dispersion).

*Changes dispersion.csv output* - Since we now only call/rank the dispersion for features above the threshold, I am reporting NA as the dispersion value for features below the threshold.  @rjrico10x, I believe this shouldn't break any tests because the number of rows reported is still the same, but if you're checking for value concordance it will.

*Fix CR-1598 and unreported errors* - Although we do the PCA on a subset of the matrix, we generate a projection matrix for the entire feature space, and insert 0's for values that weren't selected.  The logic for this was incorrect.  After thresholding and subsampling, some work is required to know which column in the smaller matrix for PCA has a given index in the larger original matrix.  I added some logic to handle this within the `get_original_columns_used` method.

*Misc*

I added a doc string to a method that confused me, and added some more warnings and error messages.

* Remove print statement

* Fix Convergence Criteria Error

smax was always being set to 1 instead of the largest singular value because when the R code was ported over, the variable `iter` was not renamed in one spot and so instead was referring to the built-in function.

* Fix Test

Attempt more numerically stable solution.

* Add Battery of Additional Tests

* Change the basic testing matrix to a structured one to ensure reproducibility
* Add tests of basic PCA and output writing
* Verify a user can reconstruct our PCs from our output files.

* Fix Message Strings '%' -> .format

* Fix string format