Fix CELLRANGER-1548 (#1191)
* fix formatting in for usable reads, median UMIs (#1190)

* Fix CELLRANGER-1548

One of the edge cases we feared has appeared, this commit is a quick low-blast-radius change to deal with this issue, a proper job should be done latter.

The PCA code makes a number of pre-processing checks on the matrix, including thresholding, downsampling, and normalization.  These checks are done serially, but later steps can invalidate the earlier check that was done, leading to :(.

CR-1548 is an example of this.  The code is passed a matrix of size 15953 x 1200. This matrix is then thresholded so that no row/column with a zero count is allowed.  Next, a random subset of barcodes, and a subset of features are selected to further reduce this matrix.  In the CR-1548 case, this takes a matrix of size 15953 x 1200 and reduces it to size 900 x 600.  As as a result, some rows/columns which had previously passed the threshold and had counts >0 no longer do.  Additionally, there are now some rows/columns with zero variance as all the cells with counts for particular genes were removed, leading to lots of NaN, Inf, etc. when scaling a column by its variance.

This is a quick fix to get around this.  We modify the `normalize_and_transpose` method to return a scaling factor of 1 in the case that the variance is zero.  This avoids problems caused by scaling rows with variance == 0.