CELLRANGER-1667 Fix Memory Estimates (#1196)
The memory estimate for the CLOUPE_PREPROCESS stage previously underestimated the amount of memory needed for the `crconverter` program, leading to some jobs being killed because they exceeded their resource quota.

This commit makes a new memory estimate, but to maintain backwards compatibility (and thus minimize the effect of a change at this stage) it takes the maximum of the either the old estimate or the new estimate.  A review of the old and new estimate is below.

**Old Estimate**

The old estimate was based on the size of the matrices in python.  However, the crconverter program is written in Go, which has a garbage collector in the runtime and performs enough additional work where the simple python estimate of the matrix size was no longer accurate.  However, for reference a simple (R) function that calculates the GB requested by the old estimate, given a sparse matrix with a certain number of barcodes (bcs) and non-zero-entries (nnz) is shown below.

    oldestimate <- function(bcs, nnz) {
      mmg = nnz / 50e6 # 20 bytes per sparse matrix entry
      mmg = mmg + bcs / 2000000 # 50 bytes per barcode
      old = 2 * 2 * round(ceiling(mmg)) # Safety factor of 4
      old = max(6, old) # never less than 6 GB
      max(3, old) # yes, 3<6, but this is in the python code
    }

**New Estimate**

Profiling showed that for this GEX data, within the `crconverter` program a significant amount of memory is consumed during the call to `ConvertFilteredMatrixH5`, in particular when the functions `NewCLoupeFeatureBarcodeMatrix` and `WriteFeatureBarcodeMatrix` are called.  Examining the Go runtime statistics (runtime::MemStats.Alloc) showed the Heap Allocated memory is not consistent across runs, as the number of GCs appears to be stochastic, resulting in a higher or lower working set across replicates.  Although a wide variety of datasets would be ideal, to build a better model I subsampled the data from CR-1667 to get matrices with various numbers of features(FEATS)/barcodes(BCS)/non-zero-entries(NNZ), this data shown is shown at the end.  For each combination, the maximum amount of virtual memory (VSIZE, from `ps -o vsize=`) was recorded as `crconverter` ran.

Regressing VSIZE agains the BCS, FEATS and NNZ gave an R2 of 0.99.  However, VSIZE was mostly colinear with NNZ and NNZ was also the only siginifcant term in the multiple regression.  As a univariate regression with just that variable led to an equivalent R2 of 0.99, I selected a simpler model with only NNZ as a predictor.  This led to the new formula below:

    GB_Needed = 1.691 + 8.537e-8 * NNZ

To create the new estimate, I take this formula, multiply it by a safety factor of 1.5.  That 50% factor is completely arbitrary, as the model is basically a perfect predictor so any errors will be due to things not accounted for by the model/training-data. 50% is a total guess for differences from unaccounted-for-factors.  After this safety factor is accounted for, I then take the rounded up integer of this.

Stepping back, it's worth noting that this new estimate effectively allocates 85 bytes per NNZ entry, whereas the old estimate allocated 20 bytes, leading to the new estimate giving higher memory allocations for matrices with large NNZ.

**Data matrix** - Data used for model development below:

**Experiment**|**NNZ**|**BCS**|**FEATURES**|**MAX.MB**|**MaxVSize (GB)**|**Old.Prediction**|**New.Prediction**
:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:|:-----:
1|8827|796|310|465|1.559372|6|2
2|9668|796|310|465|1.617608|6|2
3|5732300|19905|7763|664|2.602776|6|3
4|11855182|19905|15526|1126|2.5478|6|3
5|11681297|19905|15526|1259|2.494684|6|3
6|23335610|19905|31053|2117|3.44752|6|4
7|11425299|39811|7763|1192|2.492572|6|3
8|24004973|39811|15526|2447|3.834968|6|4
9|46686200|39811|31053|4231|5.661336|6|6
10|50007794|53080|24842|4892|6.477916|8|6
11|69960606|59716|31053|6214|7.888216|8|8
12|74731115|63697|31053|6213|7.83062|8|9
13|84472373|75640|29500|6742|8.511328|8|9
14|24085670|79622|7763|2514|4.115616|6|4
15|47454794|79622|15526|4099|5.526488|6|6
16|93334009|79622|31053|7997|9.854048|8|10