WIP FIX CELLRANGER-1512 - Reduce Memory Consumption > 2X (#1257)
* FIX CELLRANGER-1512 - Memory Consumption Reduction > 5X

Cellranger was exceeding its memory limit in the summarize_reports stage. Instrumenting the python script by inserting this line throughout the code:

   import resource
   print resource.getrusage(resource.RUSAGE_SELF).ru_maxrss

Shows that this process consumes at least 4.294 GB of memory, and so was always over the 4 GB RSS limit.  The bulk of this memory allocation occurs during the call to the function that builds the web_summary, after which the ru_maxrss value increases by an order of magnitude.  Further tracking showed the large memory increase was due to this line inside the `counter_barcode_rank_plot_data` function.

        sorted_bc, sorted_counts = zip(*sorted(zip(bc_sequence, counts_per_bc), key=lambda pair: pair[1], reverse=True))

The problem is that zip, sorted, and zip all create copies of the original data, and worse wrap the underlying primitives in expensive tuples.  Instead of this expensive operation, I changed the code to use a simpler path with fewer memory allocations.  This dropped the ru_maxrss value down from 4,292,816 to 981,444, which puts us again under the limit.

* Revert to old sort order

Use sorted instead, requires more memory in profiling (1,675,060) but still under
the limit.